Moved social preview forms to components
